+++ edu.rice.cs.hpcviewer.ui/src/edu/rice/cs/hpcviewer/ui/ProfilePart.java
@@ -35,6 +35,8 @@ import edu.rice.cs.hpcviewer.ui.graph.GraphEditorInput;
 import edu.rice.cs.hpcviewer.ui.graph.GraphHistoViewer;
 import edu.rice.cs.hpcviewer.ui.graph.GraphPlotRegularViewer;
 import edu.rice.cs.hpcviewer.ui.graph.GraphPlotSortViewer;
+import edu.rice.cs.hpcviewer.ui.internal.AbstractBaseViewItem;
+import edu.rice.cs.hpcviewer.ui.internal.AbstractViewItem;
 import edu.rice.cs.hpcviewer.ui.parts.bottomup.BottomUpView;
 import edu.rice.cs.hpcviewer.ui.parts.datacentric.Datacentric;
 import edu.rice.cs.hpcviewer.ui.parts.editor.Editor;
@@ -42,8 +44,6 @@ import edu.rice.cs.hpcviewer.ui.parts.flat.FlatView;
 import edu.rice.cs.hpcviewer.ui.parts.thread.ThreadView;
 import edu.rice.cs.hpcviewer.ui.parts.thread.ThreadViewInput;
 import edu.rice.cs.hpcviewer.ui.parts.topdown.TopDownView;
-import edu.rice.cs.hpcviewer.ui.tabItems.AbstractBaseViewItem;
-import edu.rice.cs.hpcviewer.ui.tabItems.AbstractViewItem;
