+++ edu.rice.cs.hpcviewer.ui/src/edu/rice/cs/hpcviewer/ui/internal/AbstractBaseViewItem.java
@@ -1,4 +1,4 @@
-package edu.rice.cs.hpcviewer.ui.tabItems;
+package edu.rice.cs.hpcviewer.ui.internal;
 
 import org.eclipse.swt.custom.CTabFolder;
 import org.eclipse.swt.custom.CTabItem;
