+++ edu.rice.cs.hpcviewer.ui/src/edu/rice/cs/hpcviewer/ui/parts/bottomup/BottomUpView.java
@@ -9,7 +9,7 @@ import edu.rice.cs.hpc.data.experiment.Experiment;
 import edu.rice.cs.hpc.data.experiment.scope.RootScope;
 import edu.rice.cs.hpc.data.experiment.scope.RootScopeType;
 import edu.rice.cs.hpcviewer.ui.base.IViewBuilder;
-import edu.rice.cs.hpcviewer.ui.tabItems.AbstractViewItem;
+import edu.rice.cs.hpcviewer.ui.internal.AbstractViewItem;
 
 public class BottomUpView extends AbstractViewItem {
